+++ examples/network.jl
@@ -1,20 +1,25 @@
 using Pkg
-# Activate the test-environment, where PrettyTables and HiGHS are added as dependencies.
-Pkg.activate(joinpath(@__DIR__, "../test"))
+# Activate the local environment including EnergyModelsBase, HiGHS, PrettyTables
+Pkg.activate(@__DIR__)
 # Install the dependencies.
 Pkg.instantiate()
-# Add the package EnergyModelsBase to the environment.
-Pkg.develop(path=joinpath(@__DIR__, ".."))
 
+# Import the required packages
 using EnergyModelsBase
 using JuMP
 using HiGHS
 using PrettyTables
 using TimeStruct
 
+"""
+    generate_example_data()
 
-function generate_data()
-    @info "Generate case data"
+Generate the data for an example consisting of a simple electricity network.
+The more stringent CO₂ emission in latter investment periods force the utilization of the
+more expensive natural gas power plant with CCS to reduce emissions.
+"""
+function generate_example_data()
+    @info "Generate case data - Simple network example"
 
     # Define the different resources and their emission intensity in tCO2/MWh
     NG = ResourceEmit("NG", 0.2)
@@ -32,7 +37,7 @@ function generate_data()
     # can also be extracted using the function `duration` of a `SimpleTimes` structure.
     # This implies, that a strategic period is 8 times longer than an operational period,
     # resulting in the values below as "/8h".
-    op_per_strat = duration(operational_periods)
+    op_per_strat = op_duration * op_number
 
     # Creation of the time structure and global data
     T = TwoLevel(4, 1, operational_periods; op_per_strat)
@@ -55,23 +60,23 @@ function generate_data()
     # Create the individual test nodes, corresponding to a system with an electricity demand/sink,
     # coal and nautral gas sources, coal and natural gas (with CCS) power plants and CO₂ storage.
     nodes = [
-        GenAvailability(1, products),
+        GenAvailability("Availability", products),
         RefSource(
-            2,                          # Node id
+            "NG source",                # Node id
             FixedProfile(1e12),         # Capacity in MW
             FixedProfile(30),           # Variable OPEX in EUR/MW
             FixedProfile(0),            # Fixed OPEX in EUR/8h
             Dict(NG => 1),              # Output from the Node, in this gase, NG
         ),
         RefSource(
-            3,                          # Node id
+            "coal source",              # Node id
             FixedProfile(1e12),         # Capacity in MW
             FixedProfile(9),            # Variable OPEX in EUR/MWh
             FixedProfile(0),            # Fixed OPEX in EUR/8h
             Dict(Coal => 1),            # Output from the Node, in this gase, coal
         ),
         RefNetworkNode(
-            4,                          # Node id
+            "NG+CCS power plant",       # Node id
             FixedProfile(25),           # Capacity in MW
             FixedProfile(5.5),          # Variable OPEX in EUR/MWh
             FixedProfile(0),            # Fixed OPEX in EUR/8h
@@ -82,7 +87,7 @@ function generate_data()
             [capture_data],             # Additonal data for emissions and CO₂ capture
         ),
         RefNetworkNode(
-            5,                          # Node id
+            "coal power plant",         # Node id
             FixedProfile(25),           # Capacity in MW
             FixedProfile(6),            # Variable OPEX in EUR/MWh
             FixedProfile(0),            # Fixed OPEX in EUR/8h
@@ -91,7 +96,7 @@ function generate_data()
             [emission_data],            # Additonal data for emissions
         ),
         RefStorage(
-            6,                          # Node id
+            "CO2 storage",              # Node id
             FixedProfile(60),           # Rate capacity in t/h
             FixedProfile(600),          # Storage capacity in t
             FixedProfile(9.1),          # Storage variable OPEX for the rate in EUR/t
@@ -103,7 +108,7 @@ function generate_data()
             # In practice, for CO₂ storage, this is never used.
         ),
         RefSink(
-            7,                          # Node id
+            "electricity demand",       # Node id
             OperationalProfile([20 30 40 30]), # Demand in MW
             Dict(:surplus => FixedProfile(0), :deficit => FixedProfile(1e6)),
             # Line above: Surplus and deficit penalty for the node in EUR/MWh
@@ -113,15 +118,15 @@ function generate_data()
 
     # Connect all nodes with the availability node for the overall energy/mass balance
     links = [
-        Direct(14, nodes[1], nodes[4], Linear())
-        Direct(15, nodes[1], nodes[5], Linear())
-        Direct(16, nodes[1], nodes[6], Linear())
-        Direct(17, nodes[1], nodes[7], Linear())
-        Direct(21, nodes[2], nodes[1], Linear())
-        Direct(31, nodes[3], nodes[1], Linear())
-        Direct(41, nodes[4], nodes[1], Linear())
-        Direct(51, nodes[5], nodes[1], Linear())
-        Direct(61, nodes[6], nodes[1], Linear())
+        Direct("Av-NG_pp", nodes[1], nodes[4], Linear())
+        Direct("Av-coal_pp", nodes[1], nodes[5], Linear())
+        Direct("Av-CO2_stor", nodes[1], nodes[6], Linear())
+        Direct("Av-demand", nodes[1], nodes[7], Linear())
+        Direct("NG_src-av", nodes[2], nodes[1], Linear())
+        Direct("Coal_src-av", nodes[3], nodes[1], Linear())
+        Direct("NG_pp-av", nodes[4], nodes[1], Linear())
+        Direct("Coal_pp-av", nodes[5], nodes[1], Linear())
+        Direct("CO2_stor-av", nodes[6], nodes[1], Linear())
     ]
 
     # WIP data structure
@@ -134,25 +139,26 @@ function generate_data()
     return case, model
 end
 
-# Create the case and model data and run the model
-case, model = generate_data()
+# Generate the case and model data and run the model
+case, model = generate_example_data()
 optimizer = optimizer_with_attributes(HiGHS.Optimizer, MOI.Silent() => true)
 m = run_model(case, model, optimizer)
 
 # Display some results
+ng_ccs_pp, coal_pp,  = case[:nodes][[4,5]]
 @info "Capacity usage of the coal power plant"
 pretty_table(
     JuMP.Containers.rowtable(
         value,
-        m[:cap_use][case[:nodes][5], :];
+        m[:cap_use][coal_pp, :];
         header=[:t, :Value]
     ),
 )
-@info "Capacity usage of the nautral gas + CCS power plant"
+@info "Capacity usage of the natural gas + CCS power plant"
 pretty_table(
     JuMP.Containers.rowtable(
         value,
-        m[:cap_use][case[:nodes][4], :];
+        m[:cap_use][ng_ccs_pp, :];
         header=[:t, :Value]
     ),
 )