+++ test/test_examples.jl
@@ -1,7 +1,6 @@
 
 @testset "Run examples" begin
     exdir = joinpath(@__DIR__, "../examples")
-
     files = first(walkdir(exdir))[3]
     for file in files
         if splitext(file)[2] == ".jl"
@@ -13,9 +12,5 @@
             end
         end
     end
-
-    # Cleanup the test environment. Remove EnergyModelsBase from the environment,
-    # since it is added with `Pkg.develop` by the examples. The tests can not be run with
-    # with the package in the environment.
-    Pkg.rm("EnergyModelsBase")
+    Pkg.activate(@__DIR__)
 end