+++ test/runtests.jl
@@ -10,9 +10,9 @@ const TS = TimeStruct
 const TEST_ATOL = 1e-6
 
 @testset "EnergyModelsBase" begin
-    include("example.jl")
-    include("nodes.jl")
-    include("modeltype.jl")
+    include("test_general.jl")
+    include("test_nodes.jl")
+    include("test_modeltype.jl")
     include("test_examples.jl")
     include("test_utils.jl")
 end