+++ NEWS.md
@@ -1,4 +1,10 @@
 # Release notes
+=============
+
+Unversioned
+--------------------------
+ * Fixed a bug when running the examples from a non-cloned version of `EnergyModelsBase`.
+ * This was achieved through a separate Project.toml in the examples.
 
 Version 0.6.5 (2024-01-31)
 --------------------------