+++ NEWS.md
@@ -7,6 +7,7 @@ Unversioned
  * This was achieved through a separate Project.toml in the examples.
  * Added functions `inputs`, `outputs`, and `data_nodes` for `Availability` and `outputs` for `Source` nodes.
  * Allow availability to not require all resources in the the `input` and `output` field.
+ * Moved all files declaring structures to a separate folder for improved readability.
 
 Version 0.6.5 (2024-01-31)
 --------------------------